refactor: moves service disruption to separate module
