+++ src/translations/components/ServiceDisruptions.ts
@@ -1,0 +1,18 @@
+import {translation as _} from '../commons';
+
+const ServiceDisruptionsTexts = {
+  header: {
+    title: _('Driftsavvik', 'Service disruptions'),
+  },
+  body: _(
+    'Oversikt over endringer, forsinkelser og avvik finner du på atb.no:',
+    'Get an overview of changes, delays and service disruptions at atb.no (Norwegian only):      ',
+  ),
+  button: {
+    text: _(
+      'atb.no/driftsavvik (åpnes i nettleser)      ',
+      'atb.no/driftsavvik (opens in browser)',
+    ),
+  },
+};
+export default ServiceDisruptionsTexts;
