+++ src/translations/index.ts
@@ -16,7 +16,7 @@ export {default as SectionTexts} from './components/Section';
 export {default as MapTexts} from './components/Map';
 export {default as ScreenHeaderTexts} from './components/ScreenHeader';
 export {default as BottomSheetTexts} from './components/BottomSheet';
-export {default as DisappearingHeaderTexts} from './components/DisappearingHeader';
+export {default as ServiceDisruptionsTexts} from './components/ServiceDisruptions';
 export {default as LocationSearchTexts} from './screens/subscreens/LocationSearch';
 export {default as ProductTexts} from './screens/subscreens/Product';
 export {default as TravellersTexts} from './screens/subscreens/Travellers';
